Avoid pointer issue in removing procedure
During the removing procedure, since the removing node has been swaped
and removed. The right subtree of the node is pointed by anothor node
rather than the removed one. However, the push operation was done in a
wrong way. Therefore, the subtree will be balanced via the removed
node's pointer. Since the balance procedure may update the pointer to
the root of the subtree, the subtree cannot be accessed correctly after
the balance procedure.
Using a workaround way to balance the sub tree via correctly pointer.
FIXME: Update the subtree by a better way